Calculated estimated marginal means.

This function estimates the estimated marginal means (also known as least-square means) and, if relevant, the treatment effects for mmrm, lme, and ancova models.

Usage

aba_emmeans(model)

Arguments

model

abaModel. The fitted aba model to run emmeans on.

Value

an abaEmmeans object. This object contains the emmeans, the paired comparisons (i.e., treatment effect), and the sample size at each visit.

Details

This function is based on the emmeans::emmeans function. This function will only be run for the stats which are supported by emmeans.

Examples

# process data: take first 4 visits, only MCI, use CSF abeta as "treatment",
# and create endpoint as change from baseline in cognition at each visit
df <- adnimerge %>%
  dplyr::filter(
    VISCODE %in% c('bl','m06','m12','m24'),
    !is.na(CSF_ABETA_STATUS_bl),
    DX_bl %in% c('MCI')
  ) %>%
  dplyr::mutate(
    TREATMENT = factor(CSF_ABETA_STATUS_bl, levels=c(0,1),
                       labels=c('Placebo','Treatment')),
    ADAS13 = ADAS13 - ADAS13_bl,
    CDRSB = CDRSB - CDRSB_bl,
    MMSE = MMSE - MMSE_bl
  )

# fit mmrm model for different endpoints, adjusted for covariates
model <- df %>% aba_model() %>%
  set_outcomes(CDRSB, ADAS13, MMSE) %>%
  set_covariates(
    AGE, GENDER, EDUCATION
  ) %>%
  set_stats(
    stat_mmrm(id = 'RID', time = 'VISCODE', treatment = 'TREATMENT')
  ) %>%
  aba_fit()

# run emmeans
model_emmeans <- model %>% aba_emmeans()
print(model_emmeans)
#> $emmeans
#> # A tibble: 18 × 10
#>    group outcome stat  predictor_set treatment time  estimate conf_low conf_high
#>    <chr> <chr>   <chr> <chr>         <fct>     <fct>    <dbl>    <dbl>     <dbl>
#>  1 ever… CDRSB   mmrm  M1            Placebo   m06     0.106  -0.0276     0.239 
#>  2 ever… CDRSB   mmrm  M1            Treatment m06     0.333   0.229      0.437 
#>  3 ever… CDRSB   mmrm  M1            Placebo   m12     0.161  -0.00383    0.325 
#>  4 ever… CDRSB   mmrm  M1            Treatment m12     0.607   0.480      0.735 
#>  5 ever… CDRSB   mmrm  M1            Placebo   m24     0.504   0.230      0.778 
#>  6 ever… CDRSB   mmrm  M1            Treatment m24     1.47    1.26       1.68  
#>  7 ever… ADAS13  mmrm  M1            Placebo   m06    -0.0983 -0.802      0.606 
#>  8 ever… ADAS13  mmrm  M1            Treatment m06     1.13    0.587      1.67  
#>  9 ever… ADAS13  mmrm  M1            Placebo   m12    -0.171  -0.992      0.651 
#> 10 ever… ADAS13  mmrm  M1            Treatment m12     1.24    0.608      1.87  
#> 11 ever… ADAS13  mmrm  M1            Placebo   m24     1.16    0.139      2.19  
#> 12 ever… ADAS13  mmrm  M1            Treatment m24     3.89    3.11       4.67  
#> 13 ever… MMSE    mmrm  M1            Placebo   m06    -0.167  -0.462      0.128 
#> 14 ever… MMSE    mmrm  M1            Treatment m06    -0.874  -1.10      -0.645 
#> 15 ever… MMSE    mmrm  M1            Placebo   m12    -0.127  -0.448      0.194 
#> 16 ever… MMSE    mmrm  M1            Treatment m12    -1.12   -1.37      -0.874 
#> 17 ever… MMSE    mmrm  M1            Placebo   m24    -0.468  -0.915     -0.0210
#> 18 ever… MMSE    mmrm  M1            Treatment m24    -2.17   -2.52      -1.83  
#> # … with 1 more variable: pval <dbl>
#> 
#> $pairs
#> # A tibble: 9 × 10
#>   group  outcome stat  predictor_set treatment time  estimate conf_low conf_high
#>   <chr>  <chr>   <chr> <chr>         <chr>     <fct>    <dbl>    <dbl>     <dbl>
#> 1 every… CDRSB   mmrm  M1            Placebo … m06     -0.227   -0.397   -0.0569
#> 2 every… CDRSB   mmrm  M1            Placebo … m12     -0.446   -0.656   -0.237 
#> 3 every… CDRSB   mmrm  M1            Placebo … m24     -0.968   -1.32    -0.621 
#> 4 every… ADAS13  mmrm  M1            Placebo … m06     -1.23    -2.14    -0.317 
#> 5 every… ADAS13  mmrm  M1            Placebo … m12     -1.41    -2.47    -0.347 
#> 6 every… ADAS13  mmrm  M1            Placebo … m24     -2.73    -4.05    -1.41  
#> 7 every… MMSE    mmrm  M1            Placebo … m06      0.707    0.330    1.08  
#> 8 every… MMSE    mmrm  M1            Placebo … m12      0.996    0.586    1.40  
#> 9 every… MMSE    mmrm  M1            Placebo … m24      1.70     1.13     2.27  
#> # … with 1 more variable: pval <dbl>
#> 
#> $counts
#> # A tibble: 24 × 7
#>    group      outcome stat  predictor_set treatment time      n
#>    <chr>      <chr>   <chr> <chr>         <fct>     <fct> <int>
#>  1 everyone() CDRSB   mmrm  M1            Placebo   bl      198
#>  2 everyone() ADAS13  mmrm  M1            Placebo   bl      198
#>  3 everyone() MMSE    mmrm  M1            Placebo   bl      198
#>  4 everyone() CDRSB   mmrm  M1            Placebo   m06     184
#>  5 everyone() ADAS13  mmrm  M1            Placebo   m06     183
#>  6 everyone() MMSE    mmrm  M1            Placebo   m06     186
#>  7 everyone() CDRSB   mmrm  M1            Placebo   m12     183
#>  8 everyone() ADAS13  mmrm  M1            Placebo   m12     182
#>  9 everyone() MMSE    mmrm  M1            Placebo   m12     185
#> 10 everyone() CDRSB   mmrm  M1            Placebo   m24     159
#> # … with 14 more rows
#> 
#> attr(,"class")
#> [1] "abaEmmeans"