Create an aba model.

An aba model is the foundational object in the aba package. It is composed of the following:

data: a data.frame to be used to fit the statistical models

spec: the specification for the aba model composed of the following:

groups: subsets of the data

outcomes: dependent variables in statistical fits.

covariates: independent variables which should always be included in statistical fits.

predictors: independent variables which will vary across different statistical fits.

results: the resulting fitted statistics.

Usage

aba_model(data = NULL, spec = aba_spec(), verbose = FALSE)

Arguments

data

data.frame the data to use for the object

spec

abaModelSpec the spec to use for the model. Can be created with model_spec().

verbose

logical. Whether to give a progress bar during model fitting. This can be useful if the fitting procedure is going to take a long time.

Value

An aba model which can be fitted using the aba_fit() function and which can be modified in any manner.

Examples

# use built-in data and only take the baseline visit
data <- adnimerge %>% dplyr::filter(VISCODE == 'bl')

# Create aba model w/ data, groups, outcomes, covariates, predictors, stats.
# Note that we start with piping the data into the aba_model... This is
# possible because `data` is the first argument of the `aba_model()` function
# and is useful because it gives auto-completion of variables names in Rstudio.
model <- data %>% aba_model() %>%
  set_groups(everyone(), DX_bl %in% c('MCI','AD')) %>%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) %>%
  set_covariates(AGE, GENDER, EDUCATION) %>%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) %>%
  set_stats('glm')

# get a useful view of the model spec:
print(model)
#> 
#> Groups:
#>    everyone()
#>    DX_bl %in% c("MCI", "AD")
#> 
#> Outcomes:
#>    ConvertedToAlzheimers
#>    CSF_ABETA_STATUS_bl
#> 
#> Covariates:
#>     AGE GENDER EDUCATION 
#> 
#> Predictors:
#>    PLASMA_ABETA_bl
#>    PLASMA_PTAU181_bl
#>    PLASMA_NFL_bl
#>    PLASMA_ABETA_bl + PLASMA_PTAU181_bl + PLASMA_NFL_bl
#> 
#> Stats:
#>    glm 
#>    

# model specs can be modified to build on one another and save time when
# doing sensitivity analyses. Here, we create the same model as before but
# just add APOE4 as covariate.
model2 <- model %>%
  set_covariates(AGE, GENDER, EDUCATION, APOE4)

# see this change in the model print
print(model2)
#> 
#> Groups:
#>    everyone()
#>    DX_bl %in% c("MCI", "AD")
#> 
#> Outcomes:
#>    ConvertedToAlzheimers
#>    CSF_ABETA_STATUS_bl
#> 
#> Covariates:
#>     AGE GENDER EDUCATION APOE4 
#> 
#> Predictors:
#>    PLASMA_ABETA_bl
#>    PLASMA_PTAU181_bl
#>    PLASMA_NFL_bl
#>    PLASMA_ABETA_bl + PLASMA_PTAU181_bl + PLASMA_NFL_bl
#> 
#> Stats:
#>    glm 
#>    

# Calling the `fit()` function actually triggers fitting of statistics.
model <- model %>% fit()
model2 <- model2 %>% fit()

# Access the raw results in case you care about that:
print(model$results)
#> # A tibble: 20 × 8
#>    group   outcome  stat  predictor_set predictor    covariate stat_obj stat_fit
#>    <chr>   <chr>    <chr> <chr>         <chr>        <chr>     <list>   <list>  
#>  1 "every… Convert… glm   M1            ""           AGE + GE… <abaSta… <glm>   
#>  2 "every… Convert… glm   M2            "PLASMA_ABE… AGE + GE… <abaSta… <glm>   
#>  3 "every… Convert… glm   M3            "PLASMA_PTA… AGE + GE… <abaSta… <glm>   
#>  4 "every… Convert… glm   M4            "PLASMA_NFL… AGE + GE… <abaSta… <glm>   
#>  5 "every… Convert… glm   M5            "PLASMA_ABE… AGE + GE… <abaSta… <glm>   
#>  6 "DX_bl… Convert… glm   M1            ""           AGE + GE… <abaSta… <glm>   
#>  7 "DX_bl… Convert… glm   M2            "PLASMA_ABE… AGE + GE… <abaSta… <glm>   
#>  8 "DX_bl… Convert… glm   M3            "PLASMA_PTA… AGE + GE… <abaSta… <glm>   
#>  9 "DX_bl… Convert… glm   M4            "PLASMA_NFL… AGE + GE… <abaSta… <glm>   
#> 10 "DX_bl… Convert… glm   M5            "PLASMA_ABE… AGE + GE… <abaSta… <glm>   
#> 11 "every… CSF_ABE… glm   M1            ""           AGE + GE… <abaSta… <glm>   
#> 12 "every… CSF_ABE… glm   M2            "PLASMA_ABE… AGE + GE… <abaSta… <glm>   
#> 13 "every… CSF_ABE… glm   M3            "PLASMA_PTA… AGE + GE… <abaSta… <glm>   
#> 14 "every… CSF_ABE… glm   M4            "PLASMA_NFL… AGE + GE… <abaSta… <glm>   
#> 15 "every… CSF_ABE… glm   M5            "PLASMA_ABE… AGE + GE… <abaSta… <glm>   
#> 16 "DX_bl… CSF_ABE… glm   M1            ""           AGE + GE… <abaSta… <glm>   
#> 17 "DX_bl… CSF_ABE… glm   M2            "PLASMA_ABE… AGE + GE… <abaSta… <glm>   
#> 18 "DX_bl… CSF_ABE… glm   M3            "PLASMA_PTA… AGE + GE… <abaSta… <glm>   
#> 19 "DX_bl… CSF_ABE… glm   M4            "PLASMA_NFL… AGE + GE… <abaSta… <glm>   
#> 20 "DX_bl… CSF_ABE… glm   M5            "PLASMA_ABE… AGE + GE… <abaSta… <glm>   

# Calling the `summary()` function summarises covariates and metrics in
# a useful manner
model_summary <- model %>% summary()
model2_summary <- model2 %>% summary()

# see a nicely formatted print out of the summary
print(model_summary)
#> 
#> ----------------------------------------
#> everyone() | ConvertedToAlzheimers | glm
#> ----------------------------------------
#> # A tibble: 5 × 11
#>   predictor_set AGE     GENDER   EDUCATION   PLASMA_ABETA_bl   PLASMA_PTAU181_bl
#>   <chr>         <chr>   <chr>    <chr>       <chr>             <chr>            
#> 1 M1            1.03 [… 0.70 [0… 1.02 [0.83… NA                NA               
#> 2 M2            1.03 [… 0.66 [0… 1.02 [0.84… 0.00 [0.00, 3898… NA               
#> 3 M3            1.01 [… 0.75 [0… 1.01 [0.82… NA                2.65 [1.07, 6.96…
#> 4 M4            0.95 [… 0.82 [0… 1.04 [0.84… NA                NA               
#> 5 M5            0.94 [… 0.75 [0… 1.05 [0.84… 0.00 [0.00, 6496… 2.10 [0.79, 5.90…
#> # … with 5 more variables: PLASMA_NFL_bl <chr>, AUC <chr>, AIC <chr>,
#> #   Pval <chr>, nobs <chr>
#> 
#> -------------------------------------------------------
#> DX_bl %in% c("MCI", "AD") | ConvertedToAlzheimers | glm
#> -------------------------------------------------------
#> # A tibble: 5 × 11
#>   predictor_set AGE     GENDER   EDUCATION   PLASMA_ABETA_bl   PLASMA_PTAU181_bl
#>   <chr>         <chr>   <chr>    <chr>       <chr>             <chr>            
#> 1 M1            1.06 [… 0.74 [0… 1.05 [0.83… NA                NA               
#> 2 M2            1.05 [… 0.73 [0… 1.04 [0.83… 0.00 [0.00, 4194… NA               
#> 3 M3            1.06 [… 0.76 [0… 1.08 [0.85… NA                3.56 [1.19, 12.7…
#> 4 M4            0.99 [… 0.75 [0… 1.13 [0.88… NA                NA               
#> 5 M5            0.99 [… 0.76 [0… 1.17 [0.89… 0.76 [0.00, 5783… 3.29 [1.05, 12.6…
#> # … with 5 more variables: PLASMA_NFL_bl <chr>, AUC <chr>, AIC <chr>,
#> #   Pval <chr>, nobs <chr>
#> 
#> --------------------------------------
#> everyone() | CSF_ABETA_STATUS_bl | glm
#> --------------------------------------
#> # A tibble: 5 × 11
#>   predictor_set AGE      GENDER    EDUCATION  PLASMA_ABETA_bl  PLASMA_PTAU181_bl
#>   <chr>         <chr>    <chr>     <chr>      <chr>            <chr>            
#> 1 M1            1.06 [1… 1.27 [0.… 0.99 [0.8… NA               NA               
#> 2 M2            1.03 [0… 1.01 [0.… 1.00 [0.8… 0.00 [0.00, 0.0… NA               
#> 3 M3            1.04 [0… 1.33 [0.… 0.99 [0.8… NA               3.28 [1.71, 6.93…
#> 4 M4            1.06 [1… 1.27 [0.… 0.98 [0.8… NA               NA               
#> 5 M5            1.04 [0… 1.09 [0.… 1.00 [0.8… 0.00 [0.00, 0.0… 2.64 [1.36, 5.69…
#> # … with 5 more variables: PLASMA_NFL_bl <chr>, AUC <chr>, AIC <chr>,
#> #   Pval <chr>, nobs <chr>
#> 
#> -----------------------------------------------------
#> DX_bl %in% c("MCI", "AD") | CSF_ABETA_STATUS_bl | glm
#> -----------------------------------------------------
#> # A tibble: 5 × 11
#>   predictor_set AGE      GENDER    EDUCATION  PLASMA_ABETA_bl  PLASMA_PTAU181_bl
#>   <chr>         <chr>    <chr>     <chr>      <chr>            <chr>            
#> 1 M1            1.05 [0… 2.31 [0.… 0.96 [0.7… NA               NA               
#> 2 M2            1.01 [0… 2.41 [0.… 0.96 [0.7… 0.00 [0.00, 0.0… NA               
#> 3 M3            1.05 [0… 2.46 [0.… 0.99 [0.8… NA               2.93 [1.18, 8.29…
#> 4 M4            1.04 [0… 2.29 [0.… 0.97 [0.7… NA               NA               
#> 5 M5            1.01 [0… 2.45 [0.… 0.98 [0.7… 0.00 [0.00, 0.0… 2.06 [0.80, 5.89…
#> # … with 5 more variables: PLASMA_NFL_bl <chr>, AUC <chr>, AIC <chr>,
#> #   Pval <chr>, nobs <chr>

# or access the raw summary results:
print(model_summary$results)
#> # A tibble: 184 × 10
#>    group  outcome   stat  predictor_set term   form  estimate conf_low conf_high
#>    <chr>  <chr>     <chr> <chr>         <chr>  <chr>    <dbl>    <dbl>     <dbl>
#>  1 every… Converte… glm   M1            (Inte… coef   6.83e-3 6.62e- 6    5.64e0
#>  2 every… Converte… glm   M1            AGE    coef   1.03e+0 9.62e- 1    1.11e0
#>  3 every… Converte… glm   M1            GENDER coef   7.00e-1 2.32e- 1    2.00e0
#>  4 every… Converte… glm   M1            EDUCA… coef   1.02e+0 8.32e- 1    1.26e0
#>  5 every… Converte… glm   M2            (Inte… coef   1.03e-1 1.26e- 5    8.16e2
#>  6 every… Converte… glm   M2            AGE    coef   1.03e+0 9.52e- 1    1.11e0
#>  7 every… Converte… glm   M2            GENDER coef   6.57e-1 2.15e- 1    1.90e0
#>  8 every… Converte… glm   M2            EDUCA… coef   1.02e+0 8.36e- 1    1.26e0
#>  9 every… Converte… glm   M2            PLASM… coef   1.44e-8 4.99e-26    3.90e8
#> 10 every… Converte… glm   M3            (Inte… coef   2.24e-3 2.11e- 6    1.85e0
#> # … with 174 more rows, and 1 more variable: pval <dbl>