Create a aba spec object.

An aba spec is used to create an aba model. Although it is not expected that a user will create an aba spec directly, it can still be useful to understand how aba models are defined and can also be useful for sharing model results without having to share data or fitted models. Mostly, you will want to look at the automatically generated spec from an aba model.

Usage

aba_spec(
  groups = NULL,
  outcomes = NULL,
  covariates = NULL,
  predictors = NULL,
  stats = NULL
)

Arguments

groups

vector or list of logical statements as trings. Groups are subsets of the data on which different models will be fit.

outcomes

vector or list of strings Outcomes are the dependent variables in the statistical fits.

covariates

vector of strings Covariates are independent variables which remain fixed across all statistical fits and are therefore always included with the different combinations of predictors.

predictors

vector or list of strings Predictors are independent variables which you want to vary. You can include variables on their own or in combination with others. A collection of variables is referred to as a predictor_set.

stats

string or abaStat object(s) with stat_ prefix. Stats are the actual statistical models which you want to fit on the data. Their primary functions are to 1) generate a suitable model formula given the outcome - covariate - predictor combination, and 2) to actually fit the statistical model.

Value

An abaSpec object which can be used to create an aba model and specifies how an aba model should be fit. This aba spec object can be shared to demonstrate how an aba model should be created without having to actually share any data or actual fits.

Examples

# create an aba model
model <- aba_model() %>%
  set_data(adnimerge %>% dplyr::filter(VISCODE == 'bl')) %>%
  set_groups(everyone()) %>%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) %>%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) %>%
  set_stats('glm')

# extract model spec... not super useful in its raw form
model_spec <- model$spec