Create an aba model.

An aba model is the foundational object in the aba package. It is composed of the following:

data: a data.frame to be used to fit the statistical models

spec: the specification for the aba model composed of the following:

groups: subsets of the data

outcomes: dependent variables in statistical fits.

covariates: independent variables which should always be included in statistical fits.

predictors: independent variables which will vary across different statistical fits.

results: the resulting fitted statistics.

Usage

aba_model(
  data = NULL,
  groups = NULL,
  outcomes = NULL,
  predictors = NULL,
  covariates = NULL,
  stats = NULL,
  verbose = FALSE
)

Arguments

data

data.frame the data to use for the object

groups

vector or list of logical statements as trings. Groups are subsets of the data on which different models will be fit.

outcomes

vector or list of strings Outcomes are the dependent variables in the statistical fits.

predictors

vector or list of strings Predictors are independent variables which you want to vary. You can include variables on their own or in combination with others. A collection of variables is referred to as a predictor and unique variables are referred to as a term.

covariates

vector of strings Covariates are independent variables which remain fixed across all statistical fits and are therefore always included with the different combinations of predictors.

stats

string or abaStat object(s) with stat_ prefix. Stats are the actual statistical models which you want to fit on the data. Their primary functions are to 1) generate a suitable model formula given the outcome - covariate - predictor combination, and 2) to actually fit the statistical model.

verbose

logical. Whether to give a progress bar during model fitting. This can be useful if the fitting procedure is going to take a long time.

Value

An aba model which can be fitted using the aba_fit() function and which can be modified in any manner.

Examples

# use built-in data and only take the baseline visit
data <- adnimerge %>% dplyr::filter(VISCODE == 'bl')

# Create aba model w/ data, groups, outcomes, covariates, predictors, stats.
# Note that we start with piping the data into the aba_model... This is
# possible because `data` is the first argument of the `aba_model()` function
# and is useful because it gives auto-completion of variables names in Rstudio.
model <- data %>% aba_model() %>%
  set_groups(everyone(), DX_bl %in% c('MCI','AD')) %>%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) %>%
  set_covariates(AGE, GENDER, EDUCATION) %>%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) %>%
  set_stats('glm')

# get a useful view of the model spec:
print(model)
#> 
#> Groups:
#>    G1: everyone()
#>    G2: DX_bl %in% c("MCI", "AD")
#> 
#> Outcomes:
#>    O1: ConvertedToAlzheimers
#>    O2: CSF_ABETA_STATUS_bl
#> 
#> Covariates:
#>    AGE GENDER EDUCATION
#> 
#> Predictors:
#>    P0: PLASMA_ABETA_bl
#>    P1: PLASMA_PTAU181_bl
#>    P2: PLASMA_NFL_bl
#>    P3: c("PLASMA_ABETA_bl", "PLASMA_PTAU181_bl", "PLASMA_NFL_bl")
#> 
#> Stats:
#>    glm 
#>    

# model specs can be modified to build on one another and save time when
# doing sensitivity analyses. Here, we create the same model as before but
# just add APOE4 as covariate.
model2 <- model %>%
  set_covariates(AGE, GENDER, EDUCATION, APOE4)

# see this change in the model print
print(model2)
#> 
#> Groups:
#>    G1: everyone()
#>    G2: DX_bl %in% c("MCI", "AD")
#> 
#> Outcomes:
#>    O1: ConvertedToAlzheimers
#>    O2: CSF_ABETA_STATUS_bl
#> 
#> Covariates:
#>    AGE GENDER EDUCATION APOE4
#> 
#> Predictors:
#>    P0: PLASMA_ABETA_bl
#>    P1: PLASMA_PTAU181_bl
#>    P2: PLASMA_NFL_bl
#>    P3: c("PLASMA_ABETA_bl", "PLASMA_PTAU181_bl", "PLASMA_NFL_bl")
#> 
#> Stats:
#>    glm 
#>    

# Calling the `fit()` function actually triggers fitting of statistics.
model <- model %>% fit()
model2 <- model2 %>% fit()

# Access the raw results in case you care about that:
print(model$results)
#> # A tibble: 20 × 5
#>    group outcome stat  predictor fit   
#>    <chr> <chr>   <chr> <chr>     <list>
#>  1 G1    O1      S1    Basic     <glm> 
#>  2 G1    O1      S1    P0        <glm> 
#>  3 G1    O1      S1    P1        <glm> 
#>  4 G1    O1      S1    P2        <glm> 
#>  5 G1    O1      S1    P3        <glm> 
#>  6 G1    O2      S1    Basic     <glm> 
#>  7 G1    O2      S1    P0        <glm> 
#>  8 G1    O2      S1    P1        <glm> 
#>  9 G1    O2      S1    P2        <glm> 
#> 10 G1    O2      S1    P3        <glm> 
#> 11 G2    O1      S1    Basic     <glm> 
#> 12 G2    O1      S1    P0        <glm> 
#> 13 G2    O1      S1    P1        <glm> 
#> 14 G2    O1      S1    P2        <glm> 
#> 15 G2    O1      S1    P3        <glm> 
#> 16 G2    O2      S1    Basic     <glm> 
#> 17 G2    O2      S1    P0        <glm> 
#> 18 G2    O2      S1    P1        <glm> 
#> 19 G2    O2      S1    P2        <glm> 
#> 20 G2    O2      S1    P3        <glm> 

# Calling the `summary()` function summarises covariates and metrics in
# a useful manner
model_summary <- model %>% summary()
model2_summary <- model2 %>% summary()

# see a nicely formatted print out of the summary
print(model_summary)
#> 
#> -------------------------------------
#> Group = G1 | Outcome = O1 | Stat = S1
#> -------------------------------------
#> # A tibble: 5 × 11
#>   predictor AGE     GENDER   EDUCATION   PLASMA_ABETA_bl  PLASMA_PTAU181_…
#>   <chr>     <chr>   <chr>    <chr>       <chr>            <chr>           
#> 1 Basic     1.03 [… 0.70 [0… 1.02 [0.83… NA               NA              
#> 2 P0        1.03 [… 0.66 [0… 1.02 [0.84… 0.00 [0.00, 389… NA              
#> 3 P1        1.01 [… 0.75 [0… 1.01 [0.82… NA               2.65 [1.07, 6.9…
#> 4 P2        0.95 [… 0.82 [0… 1.04 [0.84… NA               NA              
#> 5 P3        0.94 [… 0.75 [0… 1.05 [0.84… 0.00 [0.00, 649… 2.10 [0.79, 5.9…
#> # … with 5 more variables: PLASMA_NFL_bl <chr>, auc <chr>, aic <chr>,
#> #   pval <chr>, nobs <chr>
#> 
#> -------------------------------------
#> Group = G1 | Outcome = O2 | Stat = S1
#> -------------------------------------
#> # A tibble: 5 × 11
#>   predictor AGE     GENDER   EDUCATION   PLASMA_ABETA_bl  PLASMA_PTAU181_…
#>   <chr>     <chr>   <chr>    <chr>       <chr>            <chr>           
#> 1 Basic     1.06 [… 1.27 [0… 0.99 [0.86… NA               NA              
#> 2 P0        1.03 [… 1.01 [0… 1.00 [0.87… 0.00 [0.00, 0.0… NA              
#> 3 P1        1.04 [… 1.33 [0… 0.99 [0.86… NA               3.28 [1.71, 6.9…
#> 4 P2        1.06 [… 1.27 [0… 0.98 [0.86… NA               NA              
#> 5 P3        1.04 [… 1.09 [0… 1.00 [0.86… 0.00 [0.00, 0.0… 2.64 [1.36, 5.6…
#> # … with 5 more variables: PLASMA_NFL_bl <chr>, auc <chr>, aic <chr>,
#> #   pval <chr>, nobs <chr>
#> 
#> -------------------------------------
#> Group = G2 | Outcome = O1 | Stat = S1
#> -------------------------------------
#> # A tibble: 5 × 11
#>   predictor AGE     GENDER   EDUCATION  PLASMA_ABETA_bl   PLASMA_PTAU181_…
#>   <chr>     <chr>   <chr>    <chr>      <chr>             <chr>           
#> 1 Basic     1.06 [… 0.74 [0… 1.05 [0.8… NA                NA              
#> 2 P0        1.05 [… 0.73 [0… 1.04 [0.8… 0.00 [0.00, 4194… NA              
#> 3 P1        1.06 [… 0.76 [0… 1.08 [0.8… NA                3.56 [1.19, 12.…
#> 4 P2        0.99 [… 0.75 [0… 1.13 [0.8… NA                NA              
#> 5 P3        0.99 [… 0.76 [0… 1.17 [0.8… 0.76 [0.00, 5783… 3.29 [1.05, 12.…
#> # … with 5 more variables: PLASMA_NFL_bl <chr>, auc <chr>, aic <chr>,
#> #   pval <chr>, nobs <chr>
#> 
#> -------------------------------------
#> Group = G2 | Outcome = O2 | Stat = S1
#> -------------------------------------
#> # A tibble: 5 × 11
#>   predictor AGE     GENDER   EDUCATION   PLASMA_ABETA_bl  PLASMA_PTAU181_…
#>   <chr>     <chr>   <chr>    <chr>       <chr>            <chr>           
#> 1 Basic     1.05 [… 2.31 [0… 0.96 [0.79… NA               NA              
#> 2 P0        1.01 [… 2.41 [0… 0.96 [0.77… 0.00 [0.00, 0.0… NA              
#> 3 P1        1.05 [… 2.46 [0… 0.99 [0.80… NA               2.93 [1.18, 8.2…
#> 4 P2        1.04 [… 2.29 [0… 0.97 [0.79… NA               NA              
#> 5 P3        1.01 [… 2.45 [0… 0.98 [0.79… 0.00 [0.00, 0.0… 2.06 [0.80, 5.8…
#> # … with 5 more variables: PLASMA_NFL_bl <chr>, auc <chr>, aic <chr>,
#> #   pval <chr>, nobs <chr>

# or access the raw summary results:
print(model_summary$results)
#> $coefs
#> # A tibble: 84 × 9
#>    group outcome stat  predictor term    estimate conf_low conf_high  pval
#>    <chr> <chr>   <chr> <chr>     <chr>      <dbl>    <dbl>     <dbl> <dbl>
#>  1 G1    O1      S1    Basic     AGE      1.03e+0 9.62e- 1    1.11e0 0.361
#>  2 G1    O1      S1    Basic     GENDER   7.00e-1 2.32e- 1    2.00e0 0.510
#>  3 G1    O1      S1    Basic     EDUCAT…  1.02e+0 8.32e- 1    1.26e0 0.868
#>  4 G1    O1      S1    P0        AGE      1.03e+0 9.52e- 1    1.11e0 0.507
#>  5 G1    O1      S1    P0        GENDER   6.57e-1 2.15e- 1    1.90e0 0.443
#>  6 G1    O1      S1    P0        EDUCAT…  1.02e+0 8.36e- 1    1.26e0 0.833
#>  7 G1    O1      S1    P0        PLASMA…  1.44e-8 4.99e-26    3.90e8 0.368
#>  8 G1    O1      S1    P1        AGE      1.01e+0 9.41e- 1    1.09e0 0.715
#>  9 G1    O1      S1    P1        GENDER   7.52e-1 2.41e- 1    2.23e0 0.610
#> 10 G1    O1      S1    P1        EDUCAT…  1.01e+0 8.19e- 1    1.25e0 0.937
#> # … with 74 more rows
#> 
#> $metrics
#> # A tibble: 80 × 8
#>    group outcome stat  predictor term  estimate conf_low conf_high
#>    <chr> <chr>   <chr> <chr>     <chr>    <dbl>    <dbl>     <dbl>
#>  1 G1    O1      S1    Basic     auc      0.576    0.416     0.735
#>  2 G1    O1      S1    Basic     aic    123.      NA        NA    
#>  3 G1    O1      S1    Basic     pval     1       NA        NA    
#>  4 G1    O1      S1    Basic     nobs   198       NA        NA    
#>  5 G1    O1      S1    P0        auc      0.606    0.454     0.757
#>  6 G1    O1      S1    P0        aic    124.      NA        NA    
#>  7 G1    O1      S1    P0        pval     0.361   NA        NA    
#>  8 G1    O1      S1    P0        nobs   198       NA        NA    
#>  9 G1    O1      S1    P1        auc      0.701    0.591     0.812
#> 10 G1    O1      S1    P1        aic    120.      NA        NA    
#> # … with 70 more rows
#>